Ruta POST /user hecha, crea un usuario nuevo y también está agregada la funcionalidad de crear un usuario admin automaticamente la primera vez que se ejecuta el servidor
